HealthFraudMLChain/requirements.txt
# Core Flask dependencies
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2

# Production WSGI server
gunicorn==21.2.0

# Database
pymongo==4.6.1
dnspython==2.6.1

# Machine Learning
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.0
scipy==1.12.0
joblib==1.3.2

# Encryption
eciespy==0.4.1
pycryptodome==3.20.0

# Date/Time utilities
python-dateutil==2.9.0.post0
pytz==2024.1

# Environment and configuration
python-dotenv==1.0.1